build(ide): :wrench: add vscode configuration
Add Visual Studio Code configuration files (and update the .gitignore) to make it easier to use that IDE as well.

Note that .vscode/settings.json has a list of conventional commit scopes. This is NOT a definitive list! They're just the ones I could think of while quickly adding the config, and I'm sure we'll need many more as the project matures.