feat: :bento: add favicon
Add a baseline favicon. There's much more to be done (apple-touch-icon, maskable, etc.), but those will come later. The root route links to the SVG icon, browsers that don't like it should fall back to /favicon.ico. The SVG isn't linked to in the root error boundary, so it will use the .ico there as well.

The SVG is responsive to prefers-color-scheme, using some fancy media queries embedded in the SVG. That, too, I designed to be failsafe: renderers that don't support media queries should ignore the style tag and fall back to the light mode icon.