Add database (#41)
* feat(database): :construction: install prisma and add scripts

Install Prisma ORM and write some scripts to init an run Postgres. This does not include the Prisma client package, any Remix loaders or instances of Prisma, or models in the database schema.

* feat(database): :card_file_box: create database schema

Create a Prisma database schema and initial migration.

* feat(database): :card_file_box: connect Remix to Prisma

Add a server script so routes can access the database throught Prisma.

* feat(database): :card_file_box: add simple database test page

Add a counter to /events that shows the number of event records in the database. This helps verify the database is working.

* fix(database): :construction_worker: postinstall workaround for vercel database

Add `prisma generate` to package.json postinstall script, to fix issues with Prisma and Vercel's dependency caching (https://pris.ly/d/vercel-build).