feat: :passport_control: initial work on auth and login
Install & set up remix-auth, by adding Cred and Session tables to the database, installing & setting up out first strategy, remix-auth-form (with lots of backend code to check passwords, etc.). This does not include a way to create user accounts, nor proper playwright tests (some tests should fail currently).

This is not a breaking change because we don't use the database in the app yet for any purpose that would conflict with this change.